Fix bug to show the avatar of the thread's writer
Before fixing the bug, when the thread is viewd by a guest,
an error occures because the path to the avatar of the writer
of the thread is not provided. This commit solves this bug.
Of course if the writer of the thread doen't have
an avatar then a default avatar will be used.