Minor bug fix
The logic that allows the users to subscribe to a thread doesn't allow
the creator of the thread himself to subscribe to his own thread which
is a rational logic. Mistakenly the logic that makes the visual
indicator label on a locked thread to be shown was included within that
same logic of the subscribe button in that the creator of the thread
can't see that label on his locked thread anyway. This commit fixes that
bug in that it separates the code of both.